pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-notes-manager"
version = "0.1.0"
description = "Cross-platform CLI notes navigator (menu-first) backed by a Git repo of Markdown files."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Your Name"}]
dependencies = [
  "rich>=13.7.1",
  "PyYAML>=6.0.2",
  "questionary>=2.0.1",
  "GitPython>=3.1.43",
  "typing-extensions>=4.7.1; python_version<'3.11'",
  "tomli>=2.0.1; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
notes-cli = "notes_cli.app:main"

[tool.ruff]
line-length = 100
target-version = "py310"
